PDFBOX-1196: use long instead of int for object offsets as proposed by Timo Boehme
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1227500 13f79535-47bb-0310-9956-ffa450edef68

+++ preflight/src/main/java/org/apache/padaf/preflight/helpers/StreamValidationHelper.java
@@ -215,7 +215,7 @@ public class StreamValidationHelper extends AbstractValidationHelper {
 		InputStream ra = null;
 		try {
 			ra = handler.getSource().getInputStream();
-			Integer offset = (Integer) handler.getDocument().getDocument()
+			Long offset = handler.getDocument().getDocument()
 					.getXrefTable().get(new COSObjectKey(cObj));
 
 			// ---- go to the beginning of the object
